+++ src/main/java/wooteco/subway/admin/controller/LineStationController.java
@@ -3,9 +3,8 @@ package wooteco.subway.admin.controller;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import wooteco.subway.admin.domain.Station;
-import wooteco.subway.admin.dto.LineStationCreateRequest;
 import wooteco.subway.admin.dto.LineStationCreateResponse;
-import wooteco.subway.admin.dto.LineStationRequest;
+import wooteco.subway.admin.dto.request.LineStationCreateRequest;
 import wooteco.subway.admin.dto.response.StationResponse;
 import wooteco.subway.admin.service.LineService;
 import wooteco.subway.admin.service.StationService;
@@ -21,39 +20,19 @@ public class LineStationController {
 		this.stationService = stationService;
 	}
 
-	@PostMapping("/lines/stations")
-	public ResponseEntity createLineStation(@RequestBody LineStationRequest view) {
+	@PostMapping("/lines/{lineId}/stations")
+	public ResponseEntity createLineStation(@PathVariable("lineId") Long lineId, @RequestBody LineStationCreateRequest view) {
 		Station station = stationService.findBy(view.getStationId());
-		if (isRequestForFirstStation(view)) {
-			lineService.addLineStation(view.getLineId(),
-					new LineStationCreateRequest(null, station.getId(), 0, 0));
-
-			LineStationCreateResponse lineStationCreateResponse = new LineStationCreateResponse(view.getLineId(), StationResponse.of(station));
-
-			return ResponseEntity
-					.ok()
-					.body(lineStationCreateResponse);
-		}
-
-		Station preStation = stationService.findBy(view.getPreStationId());
-
-		lineService.addLineStation(view.getLineId(),
-				new LineStationCreateRequest(preStation.getId(), station.getId(), 0, 0));
-
-		LineStationCreateResponse lineStationCreateResponse = new LineStationCreateResponse(view.getLineId(), StationResponse.of(station));
+		lineService.addLineStation(lineId, view);
+		LineStationCreateResponse lineStationCreateResponse = new LineStationCreateResponse(lineId, StationResponse.of(station));
 
 		return ResponseEntity
 				.ok()
 				.body(lineStationCreateResponse);
 	}
 
-	private boolean isRequestForFirstStation(@RequestBody LineStationRequest view) {
-		return view.getPreStationId() == null;
-	}
-
 	@DeleteMapping("/lines/{lineId}/stations/{stationId}")
 	public ResponseEntity deleteLineStation(@PathVariable("lineId") Long lineId, @PathVariable("stationId") Long stationId) {
-
 		lineService.removeLineStation(lineId, stationId);
 
 		return ResponseEntity
